rename PathComparator.h -> InMemoryPathMap.h
