+++ src/Disks/ObjectStorages/InMemoryPathMap.h
